requirements.txt
Django==5.2.6
djangorestframework==3.14.0
python-decouple==3.8
python-dotenv==1.0.0
gunicorn==21.2.0
whitenoise==6.6.0
Pillow>=10.0.0
django-storages[google]==1.14.2
Google-Cloud-Storage==2.10.0requests>=2.28.0